Convert to OOP
This is just the initial OOP conversion commit. Things are pretty
different. The calcuations are the same, but access is through OOP
objects. See code for further modifications.
Note - The following files are being deprecated and will be removed
in a later commit:
Adsorption_Analysis.py
background_Adsorption_Analysis.py,
checkPass.py,
and dataStructure.py